cleanup file structure cleanup naming add IO::UDP namespace for internal (IO) UDP logic
