+++ ecal/core/CMakeLists.txt
@@ -137,60 +137,62 @@ endif()
 ######################################
 # io/udp
 ######################################
-if(ECAL_NPCAP_SUPPORT)
-set(ecal_io_udp_src_npcap 
-   src/io/udp/udp_receiver_npcap.cpp
-   src/io/udp/udp_receiver_npcap.h
-)
-endif()
-
+# io/udp/fragmentation
 set(ecal_io_udp_fragmentation_src
+    src/io/udp/fragmentation/msg_type.h
     src/io/udp/fragmentation/rcv_fragments.cpp
     src/io/udp/fragmentation/rcv_fragments.h
     src/io/udp/fragmentation/snd_fragments.cpp
     src/io/udp/fragmentation/snd_fragments.h
 )
 
-set(ecal_io_udp_protobuf_src
-    src/io/udp/protobuf/rcv_logging.cpp
-    src/io/udp/protobuf/rcv_logging.h
-    src/io/udp/protobuf/rcv_sample.cpp
-    src/io/udp/protobuf/rcv_sample.h
-    src/io/udp/protobuf/snd_logging.cpp
-    src/io/udp/protobuf/snd_logging.h
-    src/io/udp/protobuf/snd_sample.cpp
-    src/io/udp/protobuf/snd_sample.h
+# io/udp/sendreceive (npcap)
+if(ECAL_NPCAP_SUPPORT)
+set(ecal_io_udp_sendreceive_src_npcap 
+   src/io/udp/sendreceive/udp_receiver_npcap.cpp
+   src/io/udp/sendreceive/udp_receiver_npcap.h
+)
+endif()
+
+# io/udp/sendreceive
+set(ecal_io_udp_sendreceive_src
+    src/io/udp/sendreceive/udp_receiver.cpp
+    src/io/udp/sendreceive/udp_receiver.h
+    src/io/udp/sendreceive/udp_receiver_asio.cpp
+    src/io/udp/sendreceive/udp_receiver_asio.h
+    src/io/udp/sendreceive/udp_sender.cpp
+    src/io/udp/sendreceive/udp_sender.h
+    ${ecal_io_udp_sendreceive_src_npcap}
 )
 
-set(ecal_io_udp_src
-    src/io/udp/msg_type.h
-    src/io/udp/topic2mcast.h
-    src/io/udp/udp_configurations.cpp
-    src/io/udp/udp_configurations.h
-    src/io/udp/udp_init.cpp
-    src/io/udp/udp_init.h
-    src/io/udp/udp_receiver.cpp
-    src/io/udp/udp_receiver.h
-    src/io/udp/udp_receiver_asio.cpp
-    src/io/udp/udp_receiver_asio.h
-    src/io/udp/udp_sender.cpp
-    src/io/udp/udp_sender.h
-    ${ecal_io_udp_src_npcap}
-)
-
-# io/udp/linux
+# io/udp/sendreceive/linux
 if(UNIX)
-  set(ecal_io_udp_linux_src
-      src/io/udp/linux/ecal_socket_os.h
+  set(ecal_io_udp_sendreceive_linux_src
+      src/io/udp/sendreceive/linux/socket_os.h
 )
 endif()
 
-# io/udp/win32
+# io/udp/sendreceive/win32
 if (WIN32)
-  set(ecal_io_udp_win_src
-      src/io/udp/win32/ecal_socket_os.h
+  set(ecal_io_udp_sendreceive_win_src
+      src/io/udp/sendreceive/win32/socket_os.h
 )
 endif()
+
+# io/udp
+set(ecal_io_udp_src
+    src/io/udp/ecal_udp_configurations.cpp
+    src/io/udp/ecal_udp_configurations.h
+    src/io/udp/ecal_udp_logging_receiver.cpp
+    src/io/udp/ecal_udp_logging_receiver.h
+    src/io/udp/ecal_udp_logging_sender.cpp
+    src/io/udp/ecal_udp_logging_sender.h
+    src/io/udp/ecal_udp_sample_receiver.cpp
+    src/io/udp/ecal_udp_sample_receiver.h
+    src/io/udp/ecal_udp_sample_sender.cpp
+    src/io/udp/ecal_udp_sample_sender.h
+    src/io/udp/ecal_udp_topic2mcast.h
+)
 
 ######################################
 # logging
@@ -343,6 +345,8 @@ if (WIN32)
     list (APPEND
         ecal_cmn_src
         src/ecal_win_main.h
+        src/ecal_win_network.cpp
+        src/ecal_win_network.h
     )
 endif()
 
@@ -462,10 +466,10 @@ ecal_add_ecal_shared_library(${PROJECT_NAME}
     ${ecal_io_shm_linux_src}
     ${ecal_io_shm_win_src}
     ${ecal_io_udp_fragmentation_src}
-    ${ecal_io_udp_protobuf_src}
+    ${ecal_io_udp_sendreceive_src}
     ${ecal_io_udp_src}
-    ${ecal_io_udp_linux_src}
-    ${ecal_io_udp_win_src}
+    ${ecal_io_udp_sendreceive_linux_src}
+    ${ecal_io_udp_sendreceive_win_src}
     ${ecal_logging_src} 
     ${ecal_monitoring_src} 
     ${ecal_pubsub_src} 
@@ -584,8 +588,10 @@ if(NOT ${CMAKE_VERSION} VERSION_LESS "3.8.0")
     ${ecal_io_shm_linux_src}
     ${ecal_io_shm_win_src} 
     ${ecal_io_udp_fragmentation_src}
-    ${ecal_io_udp_protobuf_src}
+    ${ecal_io_udp_sendreceive_src}
     ${ecal_io_udp_src}
+    ${ecal_io_udp_sendreceive_linux_src}
+    ${ecal_io_udp_sendreceive_win_src}
     ${ecal_logging_src} 
     ${ecal_monitoring_src}
     ${ecal_pubsub_src}
