+++ ecal/core/src/ecal_globals.cpp
@@ -22,7 +22,8 @@
 **/
 
 #include "ecal_globals.h"
-#include "io/udp/udp_init.h"
+#include "ecal_win_network.h"
+
 #include "config/ecal_config_reader.h"
 
 #include <stdexcept>
@@ -44,8 +45,10 @@ namespace eCAL
     // will be set if any new module was initialized
     bool new_initialization(false);
 
-    // this is needed here for functions like "GetHostName" on windows
+#ifdef ECAL_OS_WINDOWS
+    // windows network startup
     Net::Initialize();
+#endif /* ECAL_OS_WINDOWS */
 
     /////////////////////
     // CONFIG
@@ -297,8 +300,10 @@ namespace eCAL
     log_instance                    = nullptr;
     config_instance                 = nullptr;
 
-    // last not least we close all
+#ifdef ECAL_OS_WINDOWS
+    // windows network cleanup
     Net::Finalize();
+#endif /* ECAL_OS_WINDOWS */
 
     initialized = false;
