+++ ecal/core/src/ecal_process.cpp
@@ -29,7 +29,7 @@
 #include "registration/ecal_registration_receiver.h"
 #include "ecal_globals.h"
 #include "ecal_process.h"
-#include "io/udp/udp_configurations.h"
+#include "io/udp/ecal_udp_configurations.h"
 
 #include <array>
 #include <chrono>
