+++ ecal/core/src/ecal_win_network.h
@@ -5,9 +5,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,34 +17,17 @@
  * ========================= eCAL LICENSE =================================
 */
 
+/**
+ * @brief  Win32 network initialization
+**/
 
 #pragma once
 
-#include "io/udp/udp_receiver.h"
-
-#include "config/ecal_config_reader_hlp.h"
-#include <udpcap/npcap_helpers.h>
-#include <udpcap/udpcap_socket.h>
-
 namespace eCAL
 {
-  ////////////////////////////////////////////////////////
-  // Npcap based receiver class implementation
-  ////////////////////////////////////////////////////////
-  class CUDPReceiverPcap : public CUDPReceiverImpl
+  namespace Net
   {
-  public:
-    CUDPReceiverPcap(const SReceiverAttr& attr_);
-
-    bool AddMultiCastGroup(const char* ipaddr_) override;
-    bool RemMultiCastGroup(const char* ipaddr_) override;
-
-    size_t Receive(char* buf_, size_t len_, int timeout_, ::sockaddr_in* address_ = nullptr) override;
-
-  protected:
-    bool                 m_created;
-    bool                 m_broadcast;
-    Udpcap::UdpcapSocket m_socket;
-  };
-
+    int Initialize();
+    int Finalize();
+  }
 }
