+++ ecal/core/src/io/udp/fragmentation/rcv_fragments.cpp
@@ -27,117 +27,123 @@
 
 #include <cstring>
 
-CMsgDefragmentation::CMsgDefragmentation()
-  : m_timeout(0.0)
-  , m_recv_mode(rcm_waiting)
-  , m_message_id(0)
-  , m_message_total_num(0)
-  , m_message_total_len(0)
-  , m_message_curr_num(0)
-  , m_message_curr_len(0)
+namespace IO
 {
-}
-
-CMsgDefragmentation::~CMsgDefragmentation() = default;
-
-int CMsgDefragmentation::ApplyMessage(const struct SUDPMessage& ecal_message_)
-{
-  // reset timeout
-  m_timeout = std::chrono::duration<double>(0.0);
-
-  // process current packet
-  switch(ecal_message_.header.type)
+  namespace UDP
   {
-    // new message started
-  case msg_type_header:
-    OnMessageStart(ecal_message_);
-    break;
-    // message data package
-  case msg_type_content:
-    if(m_recv_mode == rcm_reading)
+    CMsgDefragmentation::CMsgDefragmentation()
+      : m_timeout(0.0)
+      , m_recv_mode(rcm_waiting)
+      , m_message_id(0)
+      , m_message_total_num(0)
+      , m_message_total_len(0)
+      , m_message_curr_num(0)
+      , m_message_curr_len(0)
     {
-      OnMessageData(ecal_message_);
     }
-    break;
-  }
 
-  // we have a complete message in the receive buffer
-  if(m_recv_mode == rcm_completed)
-  {
-    // call complete event
-    OnMessageCompleted(std::move(m_recv_buffer));
-  }
+    CMsgDefragmentation::~CMsgDefragmentation() = default;
 
-  return(0);
-}
+    int CMsgDefragmentation::ApplyMessage(const struct SUDPMessage& ecal_message_)
+    {
+      // reset timeout
+      m_timeout = std::chrono::duration<double>(0.0);
 
-int CMsgDefragmentation::OnMessageStart(const struct SUDPMessage& ecal_message_)
-{
-  // store header info
-  m_message_id        = ecal_message_.header.id;
-  m_message_total_num = ecal_message_.header.num;
-  m_message_total_len = ecal_message_.header.len;
+      // process current packet
+      switch (ecal_message_.header.type)
+      {
+        // new message started
+      case msg_type_header:
+        OnMessageStart(ecal_message_);
+        break;
+        // message data package
+      case msg_type_content:
+        if (m_recv_mode == rcm_reading)
+        {
+          OnMessageData(ecal_message_);
+        }
+        break;
+      }
 
-  // reset current message states
-  m_message_curr_num = 0;
-  m_message_curr_len = 0;
+      // we have a complete message in the receive buffer
+      if (m_recv_mode == rcm_completed)
+      {
+        // call complete event
+        OnMessageCompleted(std::move(m_recv_buffer));
+      }
 
-  // prepare receive buffer
-  m_recv_buffer.reserve(static_cast<size_t>(m_message_total_len));
+      return(0);
+    }
 
-  // switch to reading mode
-  m_recv_mode = rcm_reading;
+    int CMsgDefragmentation::OnMessageStart(const struct SUDPMessage& ecal_message_)
+    {
+      // store header info
+      m_message_id = ecal_message_.header.id;
+      m_message_total_num = ecal_message_.header.num;
+      m_message_total_len = ecal_message_.header.len;
 
-  return(0);
-}
+      // reset current message states
+      m_message_curr_num = 0;
+      m_message_curr_len = 0;
 
-int CMsgDefragmentation::OnMessageData(const struct SUDPMessage& ecal_message_)
-{
-  // check message id
-  if(ecal_message_.header.id != m_message_id)
-  {
+      // prepare receive buffer
+      m_recv_buffer.reserve(static_cast<size_t>(m_message_total_len));
+
+      // switch to reading mode
+      m_recv_mode = rcm_reading;
+
+      return(0);
+    }
+
+    int CMsgDefragmentation::OnMessageData(const struct SUDPMessage& ecal_message_)
+    {
+      // check message id
+      if (ecal_message_.header.id != m_message_id)
+      {
 #ifndef NDEBUG
-    // log it
-    eCAL::Logging::Log(log_level_debug3, "UDP Sample OnMessageData - WRONG MESSAGE PACKET ID " + std::to_string(ecal_message_.header.id));
+        // log it
+        eCAL::Logging::Log(log_level_debug3, "UDP Sample OnMessageData - WRONG MESSAGE PACKET ID " + std::to_string(ecal_message_.header.id));
 #endif
-    m_recv_mode = rcm_aborted;
-    return(-1);
-  }
+        m_recv_mode = rcm_aborted;
+        return(-1);
+      }
 
-  // check current packet counter
-  if(ecal_message_.header.num != m_message_curr_num)
-  {
+      // check current packet counter
+      if (ecal_message_.header.num != m_message_curr_num)
+      {
 #ifndef NDEBUG
-    // log it
-    eCAL::Logging::Log(log_level_debug3, "UDP Sample OnMessageData - WRONG MESSAGE PACKET NUMBER " + std::to_string(ecal_message_.header.num) + " / " + std::to_string(m_message_curr_num));
+        // log it
+        eCAL::Logging::Log(log_level_debug3, "UDP Sample OnMessageData - WRONG MESSAGE PACKET NUMBER " + std::to_string(ecal_message_.header.num) + " / " + std::to_string(m_message_curr_num));
 #endif
-    m_recv_mode = rcm_aborted;
-    return(-1);
-  }
+        m_recv_mode = rcm_aborted;
+        return(-1);
+      }
 
-  // check current packet length
-  if(ecal_message_.header.len <= 0)
-  {
+      // check current packet length
+      if (ecal_message_.header.len <= 0)
+      {
 #ifndef NDEBUG
-    // log it
-    eCAL::Logging::Log(log_level_debug3, "UDP Sample OnMessageData - WRONG MESSAGE PACKET LENGTH " + std::to_string(ecal_message_.header.len));
+        // log it
+        eCAL::Logging::Log(log_level_debug3, "UDP Sample OnMessageData - WRONG MESSAGE PACKET LENGTH " + std::to_string(ecal_message_.header.len));
 #endif
-    m_recv_mode = rcm_aborted;
-    return(-1);
+        m_recv_mode = rcm_aborted;
+        return(-1);
+      }
+
+      // copy the message part to the receive message buffer
+      m_recv_buffer.resize(m_recv_buffer.size() + static_cast<size_t>(ecal_message_.header.len));
+      memcpy(m_recv_buffer.data() + m_recv_buffer.size() - static_cast<size_t>(ecal_message_.header.len), ecal_message_.payload, static_cast<size_t>(ecal_message_.header.len));
+
+      // increase packet counter
+      m_message_curr_num++;
+
+      // increase current length
+      m_message_curr_len += ecal_message_.header.len;
+
+      // last message packet ? -> switch to completed mode
+      if (m_message_curr_num == m_message_total_num) m_recv_mode = rcm_completed;
+
+      return(0);
+    }
   }
-
-  // copy the message part to the receive message buffer
-  m_recv_buffer.resize(m_recv_buffer.size() + static_cast<size_t>(ecal_message_.header.len));
-  memcpy(m_recv_buffer.data() + m_recv_buffer.size() - static_cast<size_t>(ecal_message_.header.len), ecal_message_.payload, static_cast<size_t>(ecal_message_.header.len));
-
-  // increase packet counter
-  m_message_curr_num++;
-
-  // increase current length
-  m_message_curr_len += ecal_message_.header.len;
-
-  // last message packet ? -> switch to completed mode
-  if(m_message_curr_num == m_message_total_num) m_recv_mode = rcm_completed;
-
-  return(0);
 }
