+++ ecal/core/src/io/udp/fragmentation/snd_fragments.cpp
@@ -28,7 +28,7 @@
 #include "ecal_process.h"
 
 #include "snd_fragments.h"
-#include "io/udp/msg_type.h"
+#include "msg_type.h"
 
 namespace
 {
@@ -49,151 +49,154 @@ namespace
   }
 }
 
-namespace eCAL
+namespace IO
 {
-  size_t CreateSampleBuffer(const std::string& sample_name_, const eCAL::pb::Sample& ecal_sample_, std::vector<char>& payload_)
+  namespace UDP
   {
-    const unsigned short sample_name_size = (unsigned short)sample_name_.size() + 1;
+    size_t CreateSampleBuffer(const std::string& sample_name_, const eCAL::pb::Sample& ecal_sample_, std::vector<char>& payload_)
+    {
+      const unsigned short sample_name_size = (unsigned short)sample_name_.size() + 1;
 #if GOOGLE_PROTOBUF_VERSION >= 3001000
-    const size_t         sample_size = ecal_sample_.ByteSizeLong();
+      const size_t         sample_size = ecal_sample_.ByteSizeLong();
 #else
-    size_t         sample_size = ecal_sample_.ByteSize();
+      size_t         sample_size = ecal_sample_.ByteSize();
 #endif
-    const size_t         data_size = sizeof(sample_name_size) + sample_name_size + sample_size;
+      const size_t         data_size = sizeof(sample_name_size) + sample_name_size + sample_size;
 
-    // create payload buffer with reserved space for first message head
-    payload_.resize(data_size + sizeof(struct SUDPMessageHead));
-    char* payload_data = payload_.data() + sizeof(struct SUDPMessageHead);
+      // create payload buffer with reserved space for first message head
+      payload_.resize(data_size + sizeof(struct SUDPMessageHead));
+      char* payload_data = payload_.data() + sizeof(struct SUDPMessageHead);
 
-    // write topic name size
-    ((unsigned short*)payload_data)[0] = sample_name_size;
-    // write topic name
-    memcpy(payload_data + sizeof(sample_name_size), sample_name_.c_str(), sample_name_size);
+      // write topic name size
+      ((unsigned short*)payload_data)[0] = sample_name_size;
+      // write topic name
+      memcpy(payload_data + sizeof(sample_name_size), sample_name_.c_str(), sample_name_size);
 
-    // write payload
-    if (ecal_sample_.SerializeWithCachedSizesToArray((google::protobuf::uint8*)payload_data + sizeof(sample_name_size) + sample_name_size))
-    {
-      return data_size;
-    }
-
-    return (0);
-  }
-
-  size_t SendFragmentedMessage(char* buf_, size_t buf_len_, long bandwidth_, TransmitCallbackT transmit_cb_)
-  {
-    if (buf_ == nullptr) return(0);
-
-    size_t sent(0);
-    size_t sent_sum(0);
-
-    int32_t total_packet_num = int32_t(buf_len_ / MSG_PAYLOAD_SIZE);
-    if (buf_len_%MSG_PAYLOAD_SIZE) total_packet_num++;
-
-    // create message header
-    struct SUDPMessageHead msg_header;
-
-    switch (total_packet_num)
-    {
-    case 1:
-    {
-      // create start packet
-      msg_header.type = msg_type_header_with_content;
-      msg_header.id = -1;  // not needed for combined header / data message
-      msg_header.num = 1;
-      msg_header.len = int32_t(buf_len_);
-
-      // copy msg_header in send buffer
-      memcpy(buf_, &msg_header, sizeof(struct SUDPMessageHead));
-
-      // send single header + data package
-      sent = transmit_cb_(buf_, sizeof(struct SUDPMessageHead) + buf_len_);
-      if (sent == 0) return(sent);
-      sent_sum += sent;
-
-#ifndef NDEBUG
-      // log it
-      //std::cout << "SendRawBuffer Packet Sent - HEADER_WITH_CONTENT (" + std::to_string(sent) + " Bytes)" << std::endl;
-#endif
-    }
-    break;
-    default:
-    {
-      // calculate bandwidth timing parameter
-      long long send_sleep_us(0);
-      if (bandwidth_ > 0)
+      // write payload
+      if (ecal_sample_.SerializeWithCachedSizesToArray((google::protobuf::uint8*)payload_data + sizeof(sample_name_size) + sample_name_size))
       {
-        send_sleep_us = MSG_BUFFER_SIZE;
-        send_sleep_us *= 1000 * 1000;
-        send_sleep_us /= bandwidth_;
+        return data_size;
       }
 
-      // create start package
-      msg_header.type = msg_type_header;
+      return (0);
+    }
+
+    size_t SendFragmentedMessage(char* buf_, size_t buf_len_, long bandwidth_, TransmitCallbackT transmit_cb_)
+    {
+      if (buf_ == nullptr) return(0);
+
+      size_t sent(0);
+      size_t sent_sum(0);
+
+      int32_t total_packet_num = int32_t(buf_len_ / MSG_PAYLOAD_SIZE);
+      if (buf_len_ % MSG_PAYLOAD_SIZE) total_packet_num++;
+
+      // create message header
+      struct SUDPMessageHead msg_header;
+
+      switch (total_packet_num)
       {
-        // create random number for message id
+      case 1:
+      {
+        // create start packet
+        msg_header.type = msg_type_header_with_content;
+        msg_header.id = -1;  // not needed for combined header / data message
+        msg_header.num = 1;
+        msg_header.len = int32_t(buf_len_);
+
+        // copy msg_header in send buffer
+        memcpy(buf_, &msg_header, sizeof(struct SUDPMessageHead));
+
+        // send single header + data package
+        sent = transmit_cb_(buf_, sizeof(struct SUDPMessageHead) + buf_len_);
+        if (sent == 0) return(sent);
+        sent_sum += sent;
+
+#ifndef NDEBUG
+        // log it
+        //std::cout << "SendRawBuffer Packet Sent - HEADER_WITH_CONTENT (" + std::to_string(sent) + " Bytes)" << std::endl;
+#endif
+      }
+      break;
+      default:
+      {
+        // calculate bandwidth timing parameter
+        long long send_sleep_us(0);
+        if (bandwidth_ > 0)
         {
-          static std::mutex xorshf96_mtx;
-          const std::lock_guard<std::mutex> lock(xorshf96_mtx);
-
-          static unsigned long x = static_cast<unsigned long>(std::chrono::duration_cast<std::chrono::nanoseconds>(
-            std::chrono::high_resolution_clock::now().time_since_epoch()).count()
-            );
-          static unsigned long y = 362436069;
-          static unsigned long z = 521288629;
-
-          msg_header.id = xorshf96(x, y, z);
+          send_sleep_us = MSG_BUFFER_SIZE;
+          send_sleep_us *= 1000 * 1000;
+          send_sleep_us /= bandwidth_;
         }
-      }
-      msg_header.num = total_packet_num;
-      msg_header.len = int32_t(buf_len_);
 
-      // send start package
-      sent = transmit_cb_(&msg_header, sizeof(struct SUDPMessageHead));
-      if (sent == 0) return(sent);
-      sent_sum += sent;
+        // create start package
+        msg_header.type = msg_type_header;
+        {
+          // create random number for message id
+          {
+            static std::mutex xorshf96_mtx;
+            const std::lock_guard<std::mutex> lock(xorshf96_mtx);
+
+            static unsigned long x = static_cast<unsigned long>(std::chrono::duration_cast<std::chrono::nanoseconds>(
+              std::chrono::high_resolution_clock::now().time_since_epoch()).count()
+              );
+            static unsigned long y = 362436069;
+            static unsigned long z = 521288629;
+
+            msg_header.id = xorshf96(x, y, z);
+          }
+        }
+        msg_header.num = total_packet_num;
+        msg_header.len = int32_t(buf_len_);
+
+        // send start package
+        sent = transmit_cb_(&msg_header, sizeof(struct SUDPMessageHead));
+        if (sent == 0) return(sent);
+        sent_sum += sent;
 
 #ifndef NDEBUG
-      // log it
-      //std::cout << "SendRawBuffer Packet Sent - HEADER (" + std::to_string(sent) + " Bytes)" << std::endl;
+        // log it
+        //std::cout << "SendRawBuffer Packet Sent - HEADER (" + std::to_string(sent) + " Bytes)" << std::endl;
 #endif
 
       // send data packages
-      msg_header.type = msg_type_content;
-      for (int32_t current_packet_num = 0; current_packet_num < total_packet_num; current_packet_num++)
-      {
-        // calculate current payload
-        size_t current_snd_len = buf_len_;
-        if (current_snd_len > MSG_PAYLOAD_SIZE) current_snd_len = MSG_PAYLOAD_SIZE;
-
-        if (current_snd_len > 0)
+        msg_header.type = msg_type_content;
+        for (int32_t current_packet_num = 0; current_packet_num < total_packet_num; current_packet_num++)
         {
-          // reduce total send len
-          buf_len_ -= current_snd_len;
+          // calculate current payload
+          size_t current_snd_len = buf_len_;
+          if (current_snd_len > MSG_PAYLOAD_SIZE) current_snd_len = MSG_PAYLOAD_SIZE;
 
-          // create data packet numbering
-          msg_header.num = current_packet_num;
-          msg_header.len = int32_t(current_snd_len);
+          if (current_snd_len > 0)
+          {
+            // reduce total send len
+            buf_len_ -= current_snd_len;
 
-          // copy msg_header in send buffer
-          memcpy(buf_ + static_cast<size_t>(current_packet_num)*MSG_PAYLOAD_SIZE, &msg_header, sizeof(struct SUDPMessageHead));
+            // create data packet numbering
+            msg_header.num = current_packet_num;
+            msg_header.len = int32_t(current_snd_len);
 
-          // send data package
-          sent = transmit_cb_(buf_ + static_cast<size_t>(current_packet_num)*MSG_PAYLOAD_SIZE, sizeof(struct SUDPMessageHead) + current_snd_len);
-          if (sent == 0) return(sent);
-          if (send_sleep_us != 0)
-            eCAL::Process::SleepFor(std::chrono::microseconds(send_sleep_us));
+            // copy msg_header in send buffer
+            memcpy(buf_ + static_cast<size_t>(current_packet_num) * MSG_PAYLOAD_SIZE, &msg_header, sizeof(struct SUDPMessageHead));
+
+            // send data package
+            sent = transmit_cb_(buf_ + static_cast<size_t>(current_packet_num) * MSG_PAYLOAD_SIZE, sizeof(struct SUDPMessageHead) + current_snd_len);
+            if (sent == 0) return(sent);
+            if (send_sleep_us != 0)
+              eCAL::Process::SleepFor(std::chrono::microseconds(send_sleep_us));
 
 #ifndef NDEBUG
-          // log it
-          //std::cout << "SendRawBuffer Packet Sent - CONTENT (" + std::to_string(sent) + " Bytes)" << std::endl;
+            // log it
+            //std::cout << "SendRawBuffer Packet Sent - CONTENT (" + std::to_string(sent) + " Bytes)" << std::endl;
 #endif
-          sent_sum += sent;
+            sent_sum += sent;
+          }
         }
       }
-    }
-    break;
-    }
+      break;
+      }
 
-    return(sent_sum);
+      return(sent_sum);
+    }
   }
 }
