+++ ecal/core/src/io/udp/fragmentation/snd_fragments.h
@@ -32,10 +32,13 @@
 #pragma warning(pop)
 #endif
 
-namespace eCAL
+namespace IO
 {
-  size_t CreateSampleBuffer(const std::string& sample_name_, const eCAL::pb::Sample& ecal_sample_, std::vector<char>& payload_);
+  namespace UDP
+  {
+    size_t CreateSampleBuffer(const std::string& sample_name_, const eCAL::pb::Sample& ecal_sample_, std::vector<char>& payload_);
 
-  using TransmitCallbackT = std::function<size_t (const void *, const size_t)>;
-  size_t SendFragmentedMessage(char* buf_, size_t buf_len_, long bandwidth_, TransmitCallbackT transmit_cb_);
+    using TransmitCallbackT = std::function<size_t(const void*, const size_t)>;
+    size_t SendFragmentedMessage(char* buf_, size_t buf_len_, long bandwidth_, TransmitCallbackT transmit_cb_);
+  }
 }
