+++ ecal/core/src/monitoring/ecal_monitoring_impl.cpp
@@ -24,7 +24,7 @@
 #include <ecal/ecal.h>
 #include <ecal/ecal_config.h>
 
-#include "io/udp/udp_configurations.h"
+#include "io/udp/ecal_udp_configurations.h"
 #include "config/ecal_config_reader_hlp.h"
 #include "ecal_monitoring_impl.h"
