+++ testing/ecal/topic2mcast_test/src/topic2mcast_test.cpp
@@ -18,7 +18,7 @@
 */
 
 #include <ecal/ecal.h>
-#include "io/udp/topic2mcast.h"
+#include "io/udp/ecal_udp_topic2mcast.h"
 
 #include <string>
 #include <chrono>
